requirements.txt
streamlit
fastapi
uvicorn==0.30.1
google-generativeai==0.5.2
mcp-agent
